Bugfix/PS-814 terminations w o badge

Approved-by: Bob Herrmann

+++ src/services/src/Demoulas.ProfitSharing.Services/Internal/ServiceDto/MemberSlice.cs
@@ -6,7 +6,7 @@
 /// </summary>
 internal sealed record MemberSlice
 {
-    internal int PsnSuffix { get; init; }
+    internal short PsnSuffix { get; init; }
     internal required int BadgeNumber { get; init; }
     internal int Ssn { get; init; }
     internal decimal HoursCurrentYear { get; init; }
